Move GQL actions to ItemListFeed. Wait until user data is present to render ItemListFeed
